# backend/docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Retoucher API
  version: 1.0.0-phase1
  description: |2

    Retoucher backend API.

    This API manages the full photo retouching workflow:

    - JWT authentication
    - Order creation and submission
    - Image upload
    - Editor assignment
    - Editor work and delivery
    - Supervisor review and QC
    - Client review and revision requests
    - Settlement and payment workflow
    - Order comments, threaded replies, rich annotations
    - Comment resolve/unresolve workflow
    - Activity logs
    - Notifications
    - Dashboard and summary endpoints
paths:
...
  /api/payments/settlement/{id}/detail/:
    get:
      operationId: payments_settlement_detail_retrieve
      description: جزئیات یک سفارش برای پنل settlement
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/payments/settlement/{id}/set-price/:
    post:
      operationId: payments_settlement_set_price_create
      description: |-
        ادمین مبلغ نهایی سفارش را تنظیم می‌کند
        فقط اگر هنوز escrow_held=False باشد
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/payments/settlement/{id}/settle/:
    post:
      operationId: payments_settlement_settle_create
      description: |-
        تسویه نهایی سفارش:
        - کسر از کیف‌پول مشتری
        - کمیسیون سایت
        - واریز به ادیتور
        - تغییر وضعیت به PAID
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/payments/settlement/all/:
    get:
      operationId: payments_settlement_all_retrieve
      description: همه سفارش‌ها با فیلتر وضعیت
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/payments/settlement/pending/:
    get:
      operationId: payments_settlement_pending_retrieve
      description: |-
        لیست سفارش‌های در انتظار تسویه
        قابل فیلتر بر اساس تاریخ و مبلغ
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/payments/settlement/summary/:
    get:
      operationId: payments_settlement_summary_retrieve
      description: خلاصه آماری settlement برای داشبورد ادمین
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/payments/wallet/admin/payments/pending/:
    get:
      operationId: payments_wallet_admin_payments_pending_retrieve
      description: پرداخت‌های در انتظار — نیاز به پیگیری
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/payments/wallet/me/payments/{payment_id}/:
    get:
      operationId: payments_wallet_me_payments_retrieve_2
      description: جزئیات یک پرداخت آنلاین خودم
      parameters:
      - in: path
        name: payment_id
        schema:
          type: string
          pattern: ^[0-9]+$
        required: true
      tags:
      - payments
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...